IMPALA-6885: handle empty s3 dirs in recover_partitions test
Two tests (test_duplicate_partitions, test_support_all_types)
added only directories, which is a no-op on s3. This results
in s3 test failures.

Another test (test_recover_partitions) falsely passed for s3
since its test for a malformed partition name is ambiguous
(did it pass because recovering correctly excluded or because
the directory does not exist, as in the case of s3?).

A recent change fixed the asserts in this test, so these tests
likely were falsely reporting success.

This change bundles directory creation with adding to it a new
file, which is the common case in this test. As a result, the
peculiarity with s3 is not handled via a special case each time
the filesystem is modified for new partitions.

This change also adds an explicit test for empty partition
directories, since the first change leaves this case untested.

Testing:
- tested against a local minicluster and s3

Change-Id: Ic6c74f4f2466b5870322bdf373e8a6a4d1d679f4
Reviewed-on: http://gerrit.cloudera.org:8080/10210
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>