IMPALA-7104: test_bloom_wait_time failing with timeout on asan
test_bloom_wait_time checks that its query runs in less than a certain
amount of time, to ensure we aren't waiting on bloom filters that will
never arrive.

The test sometimes goes slightly over the timeout and fails on ASAN.
This patch uses a specific_build_type_timeout to extend the allowed
timeout for ASAN.

It also fixes a bug in a similar test where we were comparing a
duration in seconds to a timeout in milliseconds, and renames some
variables to make it clearer why this was a bug.

Change-Id: I0271eb83e65b3b5f17e32f6501032f0d12dde38e
Reviewed-on: http://gerrit.cloudera.org:8080/10565
Reviewed-by: Michael Brown <mikeb@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>