IMPALA-6394: Disable dfs.namenode.replication.considerLoad
By default namenode excludes heavily-loaded datanodes from being
considered for block placement. Impala tests rely on every block
to be fully-replicated so this option should be set to false.
This is originally from IMPALA-7004 which was for 3.x branch only. The
option has been renamed to "dfs.namenode.redundancy.considerLoad" in
Hadooop 3.0.

Change-Id: I00010107cbd61bffa38ed8af4395b3a0e3d13c13
Reviewed-on: http://gerrit.cloudera.org:8080/10664
Reviewed-by: Tianyi Wang <twang@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>